move plan logic out of connectors into builds
