+++ dcpy/builds/plan.py
@@ -1,0 +1,213 @@
+import csv
+from enum import Enum
+import os
+from pathlib import Path
+from pydantic import BaseModel
+from typing import List
+import yaml
+
+from dcpy.utils import postgres
+from dcpy.utils import versions
+from dcpy.utils.logging import logger
+
+from dcpy.connectors.edm import recipes, publishing
+
+BUILD_SCHEMA = os.environ.get("BUILD_ENGINE_SCHEMA", postgres.DEFAULT_POSTGRES_SCHEMA)
+
+
+class RecipeInputsVersionStrategy(str, Enum):
+    find_latest = "find_latest"
+    copy_latest_release = "copy_latest_release"
+
+
+class VersionStrategy(str, Enum):
+    bump_latest_release = "bump_latest_release"
+    first_of_month = "first_of_month"
+
+
+class DataPreprocessor(BaseModel, use_enum_values=True, extra="forbid"):
+    module: str
+    function: str
+
+
+class InputDataset(BaseModel, use_enum_values=True, extra="forbid"):
+    name: str
+    version: str | None = None
+    file_type: recipes.DatasetType | None = None
+    version_env_var: str | None = None
+    import_as: str | None = None
+    preprocessor: DataPreprocessor | None = None
+
+    @property
+    def is_resolved(self):
+        return self.version is not None and self.version != "latest"
+
+    @property
+    def dataset(self):
+        if self.version is None or self.file_type is None:
+            raise Exception(f"Dataset {self.name} requires both version and file_type")
+
+        return recipes.Dataset(
+            name=self.name, version=self.version, file_type=self.file_type
+        )
+
+
+class InputDatasetDefaults(BaseModel, use_enum_values=True):
+    file_type: recipes.DatasetType | None = None
+    preprocessor: DataPreprocessor | None = None
+
+
+class RecipeInputs(BaseModel, use_enum_values=True):
+    missing_versions_strategy: RecipeInputsVersionStrategy | None = None
+    datasets: List[InputDataset] = []
+    dataset_defaults: InputDatasetDefaults | None = None
+
+
+class DatasetVersionType(str, Enum):
+    major = "major"
+    minor = "minor"
+
+
+class Recipe(BaseModel, use_enum_values=True, extra="forbid"):
+    name: str
+    product: str
+    base_recipe: str | None = None
+    version_type: DatasetVersionType | None = None
+    version_strategy: VersionStrategy | None = None
+    version: str | None = None
+    inputs: RecipeInputs
+
+    def is_resolved(self):
+        return self.version is not None and (
+            len(self.inputs.datasets) == 0
+            or len([x for x in self.inputs.datasets if not x.is_resolved()]) == 0
+        )
+
+
+def plan_recipe(recipe_path: Path) -> Recipe:
+    """Plan recipe versions for a product.
+
+    Similar to pip freeze, determines recipe versions to use for a build.
+    A base_recipe may be specified, in which case it's important to note that
+    the missing versions strategy will be applied AFTER the recipe inputs are
+    merged with the base.
+    """
+    recipe: Recipe = recipe_from_yaml(recipe_path)
+
+    # Determine the recipe version
+    if recipe.version is None and recipe.version_strategy is not None:
+        if recipe.version_strategy == VersionStrategy.bump_latest_release:
+            if recipe.version_type is None:
+                raise Exception("Recipe needs a 'version_type' to bump")
+            prev_version = publishing.get_latest_version(recipe.product)
+            recipe.version = versions.bump(
+                prev_version, bumped_part=recipe.version_type
+            )
+        if recipe.version_strategy == VersionStrategy.first_of_month:
+            recipe.version = versions.first_of_month()
+
+    # merge in base recipe inputs
+    base_recipe = (
+        recipe_from_yaml(recipe_path.parent / recipe.base_recipe)
+        if recipe.base_recipe is not None
+        else None
+    )
+
+    input_dataset_names = {d.name for d in recipe.inputs.datasets}
+    if base_recipe is not None:
+        for base_ds in base_recipe.inputs.datasets:
+            if base_ds.name not in input_dataset_names:
+                recipe.inputs.datasets.append(base_ds)
+
+    # Fill in omitted versions
+    previous_versions = {}
+    if (
+        recipe.inputs.missing_versions_strategy
+        == RecipeInputsVersionStrategy.copy_latest_release
+    ):
+        previous_versions = publishing.get_source_data_versions(
+            publishing.PublishKey(recipe.product, "latest")
+        ).to_dict()["version"]
+
+    for ds in recipe.inputs.datasets:
+        if ds.version is None:
+            if ds.version_env_var is not None:
+                version = os.getenv(ds.version_env_var)
+                if version is None:
+                    raise Exception(
+                        f"Dataset {ds.name} requires version env var: {ds.version_env_var}"
+                    )
+                ds.version = version
+            elif (
+                recipe.inputs.missing_versions_strategy
+                == RecipeInputsVersionStrategy.copy_latest_release
+            ):
+                ds.version = previous_versions[ds.name]
+            else:
+                ds.version = "latest"
+
+        if ds.version == "latest":
+            ds.version = recipes.get_config(ds.name, "latest")["dataset"]["version"]
+
+    return recipe
+
+
+def get_source_data_versions(recipe: Recipe):
+    """Get source data versions table in form of [schema_name, v]."""
+    return [["schema_name", "v"]] + [
+        [d.name, d.version] for d in recipe.inputs.datasets
+    ]
+
+
+def _apply_recipe_defaults(recipe: Recipe):
+    recipe.inputs.dataset_defaults = (
+        recipe.inputs.dataset_defaults
+        or InputDatasetDefaults(file_type=recipes.DatasetType.pg_dump)
+    )
+
+    for ds in recipe.inputs.datasets:
+        ds.preprocessor = ds.preprocessor or recipe.inputs.dataset_defaults.preprocessor
+        ds.file_type = ds.file_type or recipe.inputs.dataset_defaults.file_type
+
+
+def recipe_from_yaml(path: Path) -> Recipe:
+    """Import a recipe file from yaml, and validate schema."""
+    with open(path, "r", encoding="utf-8") as f:
+        s = yaml.safe_load(f)
+        recipe = Recipe(**s)
+        _apply_recipe_defaults(recipe)
+        return recipe
+
+
+def plan(recipe_file: Path) -> Path:
+    logger.info("Planning recipe")
+    lock_file = recipe_file.parent / f"{recipe_file.stem}.lock.yml"
+
+    recipe = plan_recipe(recipe_file)
+
+    with open(lock_file, "w", encoding="utf-8") as f:
+        logger.info(f"Writing recipe lockfile to {str(lock_file.absolute())}")
+        yaml.dump(recipe.model_dump(), f)
+
+    return lock_file
+
+
+def write_source_data_versions(recipe_file: Path):
+    recipe = recipe_from_yaml(recipe_file)
+    source_data_versions_path = recipe_file.parent / "source_data_versions.csv"
+    logger.info(f"Writing source data versions to {source_data_versions_path}")
+
+    sdv = get_source_data_versions(recipe)
+    unresolved_versions = [[k, v] for k, v in sdv if v == "latest"]
+    if len(unresolved_versions) > 0:
+        exception = (
+            "Recipe has unresolved versions! Can't write source "
+            + f"data versions {unresolved_versions}"
+        )
+        logger.error(exception)
+        raise Exception(exception)
+
+    with open(source_data_versions_path, "w", newline="") as csvfile:
+        writer = csv.writer(csvfile)
+        for key, value in sdv:
+            writer.writerow([key, value])
